Add files via upload
Problem Statement : 
The objective of this project is to design and develop a typing speed tester application that allows users to practice their typing skills and receive feedback on their typing speed and accuracy. The application should present users with random text for typing and calculate their typing speed in words per minute (WPM).